Rename component files, add test for labeled.
