+++ src/ts/explorer/active-state-component.tsx
