+++ src/ts/explorer/labeled/labeled-component.tsx
